requirements.txt
discord.py
motor
pymongo
python-dotenv
aiohttp
asyncio
numpy
scipy
requests
dnspython
python-dotenv
